+++ .gitignore
@@ -1,1 +1,5 @@
 .idea/workspace.xml
+.idea/
+*.pyc
+shape.memory
+
