+++ wrapper/components/server_setup.py
@@ -1,0 +1,3 @@
+class Server:
+    def __init__(self):
+        pass
