--- wrapper/configuration.dk
@@ -1,1 +1,0 @@
-{"node1": {"url": "tcp://192.168.99.103:2376", "ip": "192.168.99.103", "driver": "virtualbox", "role": "manager", "swarm": true}, "master": {"url": "tcp://192.168.99.103:2376", "ip": "192.168.99.103", "driver": "virtualbox", "role": "manager", "swarm": true}, "node2": {"url": "tcp://192.168.99.103:2376", "ip": "192.168.99.103", "driver": "virtualbox", "role": "manager", "swarm": true}, "worker": {"url": "tcp://192.168.99.103:2376", "ip": "192.168.99.103", "driver": "virtualbox", "role": "manager", "swarm": true}}
