+++ wrapper/modules/__init__.py
@@ -1,0 +1,3 @@
+import docker
+
+client = docker.from_env()
