--- wrapper/shape.memory
@@ -1,1 +1,0 @@
-{"node1": {"url": "tcp://192.168.99.105:2376", "ip": "192.168.99.105", "driver": "virtualbox", "swarm": true, "init": false, "role": "slave"}, "master": {"url": "tcp://192.168.99.100:2376", "ip": "192.168.99.100", "driver": "virtualbox", "swarm": true, "init": true, "role": "manager"}, "node2": {"url": "tcp://192.168.99.106:2376", "ip": "192.168.99.106", "driver": "virtualbox", "swarm": false, "init": false, "role": "slave"}, "worker": {"url": "tcp://192.168.99.101:2376", "ip": "192.168.99.101", "driver": "virtualbox", "swarm": true, "init": false, "role": "slave"}}
